refactor secret token checker
